Add Stage Accumulables to the accumulable objects for Q tool
